requirements.txt
streamlit
psycopg2-binary>=2.9.9
openai>=1.0.0
requests>=2.31.0
schedule>=1.2.0
python-dotenv>=1.0.0
